backend/requirements.txt
# Taxly CRM - PostgreSQL Edition
# Remove: motor, pymongo, bson (MongoDB drivers)
# Add: asyncpg, SQLAlchemy (async)

fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic[email]

# Database - PostgreSQL (production) + SQLite (local dev fallback)
sqlalchemy==2.0.46
asyncpg==0.31.0
aiosqlite==0.20.0

# Auth
bcrypt==4.1.3
httpx==0.28.1

# Excel / PDF
openpyxl==3.1.5
reportlab==4.4.9

# Misc
python-multipart==0.0.21
python-dateutil==2.9.0.post0
starlette==0.37.2
aiofiles==23.2.1
google-auth==2.29.0
